+++ lib/dbviewer/engine.rb
@@ -6,6 +6,11 @@ module Dbviewer
     config.autoload_paths << File.expand_path("../../", __FILE__)
     config.eager_load_paths << File.expand_path("../../", __FILE__)
 
+    # Register generators
+    config.app_generators do |g|
+      g.templates.unshift File.expand_path("../../generators/dbviewer/templates", __dir__)
+    end
+
     # Initialize the engine safely
     initializer "dbviewer.setup", after: :load_config_initializers do |app|
       Dbviewer.init